Refactor: move layout files
