--- src/@types/tailwind-scrollbar-hide.d.ts
@@ -1,5 +1,0 @@
-declare module "tailwind-scrollbar-hide" {
-  import type { Plugin } from "tailwindcss";
-  const scrollbarHide: Plugin;
-  export default scrollbarHide;
-}
